requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
alpaca-trade-api>=3.0.0
plotly>=5.14.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
